Clean code: restore navigation, move code related to writing new question from ReadingPageVM to WriteQuestionVM
