+++ RightToAskClient/RightToAskClient/RightToAskClient/ViewModels/WriteQuestionViewModel.cs
@@ -3,6 +3,9 @@ using System.Text.RegularExpressions;
 using RightToAskClient.Helpers;
 using RightToAskClient.Models;
 using RightToAskClient.Resx;
+using RightToAskClient.Views;
+using RightToAskClient.Views.Popups;
+using Xamarin.CommunityToolkit.Extensions;
 using Xamarin.CommunityToolkit.ObjectModel;
 using Xamarin.Forms;
 
@@ -50,6 +53,16 @@ namespace RightToAskClient.ViewModels
             set => SetProperty(ref _headerContent, value);
         }
         
+        private string _draftQuestion = "";
+        public string DraftQuestion
+        {
+            get => _draftQuestion;
+            set => SetProperty(ref _draftQuestion, value);
+        }
+        
+        // Commands
+        public IAsyncCommand ProceedButtonCommand { get; }
+        
         public WriteQuestionViewModel() : base(false, false)
         {
             BackCommand = new AsyncCommand(async () =>
@@ -82,10 +95,16 @@ namespace RightToAskClient.ViewModels
                     }
                 }
             });
+            
+            ProceedButtonCommand = new AsyncCommand(async () =>
+            {
+                OnSaveButtonClicked();
+            });
         }
 
         public void RequestUpdate(String query)
         {
+            DraftQuestion = query;
             ShowHeader = true;
             if (query.Length == 0)
             {
@@ -108,7 +127,25 @@ namespace RightToAskClient.ViewModels
             FilterChoices.SearchKeyword = query;
             hasQuery = true;
             RefreshCommand.ExecuteAsync();
-            
+        }
+        private async void OnSaveButtonClicked()
+        {
+            // Set up new question in preparation for upload. 
+            // The filters are new empty filters. 
+            var newQuestion = new Question()
+            {
+                QuestionText = DraftQuestion,
+                QuestionSuggester = (IndividualParticipant.getInstance().ProfileData.RegistrationInfo.IsRegistered)
+                    ? IndividualParticipant.getInstance().ProfileData.RegistrationInfo.uid
+                    : "",
+                Filters = new FilterChoices()
+            };
+
+            QuestionViewModel.Instance.Question = newQuestion;
+            QuestionViewModel.Instance.IsNewQuestion = true;
+
+            // instead of going to the questionDetailsPage we now go to a Background page and then a metadata page before the details page
+            await Shell.Current.GoToAsync($"{nameof(QuestionBackgroundPage)}");
         }
     }
 }
